bspwm: start pa-applet (temporary)
Not sure I want to keep this in the long run.
I needed a quick solution for my sound so I used it and just start it
from bspwm but I don't like the idea to couple the program to start with
the window manager.
I think I will look for a more generic solution later on so that it can
be shared between different window manager.