bspwm: restore the kill window hotkey
I removed it because I didn't want to be able to easly kill something.
It's an anction I prefere to do by really thinking about it and a
hotkey is to easy to access to let you the time for that.

But I had a buggy window that didn't want to close normally and I
realised it was really handy to be able to kill it that way if needed.
I will have to make sure I don't abuse this hotkey.