picom: add basic configuration but disable it by default
It's only sugar and almost tacky, don't really see the point but wanted
to try it.
Mostly I'm affraid it will only be hard on my eyes to read on a
transparent background.

So it's there but disabled, also it consume 3% of my ram on my old
laptop, it's something I my want to configure on a per hostname machine.