.profile: put ~/.local/bin first in the PATH
This will allow me to override some commands.
Let's says that I want to always use the same colors for dmenu: I can
create a ~/.local/dmenu which forwards arguements to the real dmenu but
with sensitive colors defined.
I could also decide to use rofi instead and this will allow me to use
rofi as dmenu and make it work with tools that used dmenu (as long as
they load my PATH of course).