rofi: use rofi instead of dmenu
Rofi is configuratble that's as simple as that.
I don't like suckless tools, for me they are more suck than less.
The only reason I first went toward dmenu was because it was used by a
lot of people and other tools.
But rofi -dmenu option and the rofi-dmenu package allow to use rofi
instead of dmenu out of the box.
Guess what, no need for a patch to have that :D