feh: use it instead of nitrogen
Feh can also be used as an image viewer, so yes it does more than one
thing but in this case I don't really mind.
Wallpaper is really a niche thing, I just configure it as a practice
exercice than anything else since I never see it anyway.
It's smaller than nitrogen anyway and allows to be able to open images
in ranger too.