bin: add scripts to logout, reboot and shutdown
The goal is to be able to do all those actions from a run menu with a
single word instead of having to type the command every time.
It can also be used in script like the polybar powermenu which makes it
behave the same way no matter how I ask about it.
An improvement would be to add a prompt to ask the user to confirm with
an option to force/not be interactive.