startup: disable udiskie
It take 2% of my memory, on my old laptop I don't even have 2G fo RAM so
I want to spare as much as possible of it.
Udisikie was only used to auto mount and show notifications about
devices being pluged in/out our mount/unmount.
I can easly mount/unmount with my `udisks-menu`, loosing the auto mount
is not that a big deal since I not using flash drives that often.
For the notifications, it's not that usefull anyway and if I really want
then I can add it to my `udisks-menu` script.

To be really conservative with my resources I should even get rid of
`udisks` but it means that I will have to deal with the mount options
manually, update udev rules, etc.
Which is not as simple as deploying my dotfiles, it requires updating
files in `/etc` so it add a extra deploying step.
Which is was I try to avoid as much as possible, the idea would be to
only have to deploy my dotfiles and install packages from a file
packages list in order to have my system up and running.