fzf: add default config to reverse the layout
The bottom/up layout feel uncomfortable because when I want to select
multiple entries with Tab, the cursor go to the line under the one I
just selected.
So I end up doing something like Tab > Up > Up > Tab > ...
Which is really annoying, it feels a lot more natural to have a top to
bottom layout.