vim: re-enable plugins
I disabled a bunch of them trying to understand why my neovim was so
slow on my laptop.
At the end of the day it was because the hard drive is too slow and
having to much files to load slow down neovim start process.

I still re-enable them because I don't want my dotfiles to be impacted
by a localized issue from my laptop.