ssh: set up the ssh-agent
Use a user service with systemd to start only one SSH agent at a time.

Register keys with a default name with `ssh-add` started in
`~/.shell/login`.

The goal is to only add the key once on the login so that they are
accessible during the session.

Also add a GTK 2 askpass to be able to enter any required passphrase.
This means that if one of the loaded key has a passphrase I'll be asked
to enter right after login in the session.
There is way around that but it's a bit less secure. For now it will do
the trick :)