shell: add ~/.profile to load env vars with LightDM
If I'm not mistaken LightDM will start the session using a bourn shell.

Meaning the shell which will start bspwm and the `startup` script will be
executed without `~/.zshenv` or `~/.zprofile` or `~/.zshrc` being
loaded.

A bourne shell will load the `~/.profile` file only, so I added in it
the definition of environment variables to make sure they are available.