vim: don't use the clipboard by default
At the end I'm used to have different registers and I don't mind having
to type `"+` or `"*` before copying/pasting.

An example of use case is that I copy something from another terminal,
it's now in the `+` register.
Then I want to paste it in my current buffer but not before modifying it
a bit so I might replace the paragraph to change `cip` and type
something before pasting.
If Vim uses the `+` register by default it would have erase the code I
copied from the other terminal.